requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
python-dotenv>=1.0
supabase>=2.4
pydantic>=2.6
httpx>=0.27
openai>=1.52
numpy>=2.3
